eatgo-api/src/test/java/kr/ardent/eatgo/application/MenuItemServiceTests.java
package kr.ardent.eatgo.application;

import kr.ardent.eatgo.domain.MenuItem;
import kr.ardent.eatgo.domain.MenuItemRepository;
import org.junit.Before;
import org.junit.Test;
import org.mockito.ArgumentMatchers;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.boot.test.mock.mockito.MockBean;

import java.util.ArrayList;
import java.util.List;

import static org.junit.Assert.*;
import static org.mockito.ArgumentMatchers.eq;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;

public class MenuItemServiceTests {

    private MenuItemService menuItemService;

    @Mock
    private MenuItemRepository menuItemRepository;

    @Before
    public void setUp(){
        MockitoAnnotations.initMocks(this);
        menuItemService = new MenuItemService(menuItemRepository);
    }

    @Test
    public void bulkUpdate(){
        List<MenuItem> menuItems = new ArrayList<MenuItem>();

        menuItems.add(MenuItem.builder()
        .name("kimchi")
        .name("bab")
        .build());

        menuItems.add(MenuItem.builder()
                .name("kimchi1")
                .name("bab1")
                .build());

        menuItems.add(MenuItem.builder().id(1004L).destroy(true).build());

        menuItemService.bulkUpdate(1L, menuItems);

        verify(menuItemRepository, times(2)).save(ArgumentMatchers.any());
        verify(menuItemRepository, times(1)).deleteById(eq(1004L));
    }

}